+++ src/Services/services.js
@@ -1,0 +1,5 @@
+import axios from "axios"
+
+export const getAllVideos = async() => {
+    return axios.get("/api/videos")
+}
