+++ src/backend/db/categories.js
@@ -7,7 +7,7 @@ import { v4 as uuid } from "uuid";
 
 export const categories = [
   {
-    _id: 1,
+    _id: uuid(),
     categoryName: "Computer Programming",
     description:
       "Computer programming is the process of designing and building an executable computer program to accomplish a specific computing",
