+++ src/backend/db/videos.js
@@ -2,80 +2,80 @@
  * Videos Database can be added here.
  * You can add videos of your wish with different attributes
  * */
-
+import { v4 as uuid } from "uuid";
 export const videos = [
   {
-    _id: 1,
+    _id: uuid(),
     title: "The 'Cow' Explosion: Black hole eats white Dwarf",
     creator: "NASA Jet Propulsion Laboratory",
     category: "Astrophysics",
     img_url: "https://res.cloudinary.com/debanftke/image/upload/v1648905905/AP-2_pbpxdv.webp"
   },
   {
-    _id: 2,
+    _id: uuid(),
     title: "The Cosmic Scale",
     creator: "SEA",
     category: "Cosmic Mysteries",
     img_url: "https://res.cloudinary.com/debanftke/image/upload/v1648911862/C-2_gel7rw.webp"
   },
   {
-    _id: 3,
+    _id: uuid(),
     title: "The fabric of Cosmos - What is Space? ",
     creator: "Advexon TV",
     category: "Cosmic Mysteries",
     img_url: "https://res.cloudinary.com/debanftke/image/upload/v1648911424/C-1_espffy.webp"
   },
   {
-    _id: 4,
+    _id: uuid(),
     title: "Birth of the Universe",
     creator: "Naked Science",
     category: "Universe",
     img_url: "https://res.cloudinary.com/debanftke/image/upload/v1648912024/U-2_iyks7w.webp"
   },
   {
-    _id: 5,
+    _id: uuid(),
     title: "Journey Through the Universe",
     creator: "Advexon Science Network",
     category: "Universe",
     img_url: "https://res.cloudinary.com/debanftke/image/upload/v1648912019/U-1_oar7iu.webp"
   },
   {
-    _id: 6,
+    _id: uuid(),
     title: "Underwater 'white smoker' vents",
     creator: "NASA Jet Propulsion Laboratory",
     category: "Astrobiology",
     img_url: "https://res.cloudinary.com/debanftke/image/upload/v1648911033/AB-2_sustxg.webp"
   },
   {
-    _id: 7,
+    _id: uuid(),
     title: "The Search for another Earth",
     creator: "NASA Jet Propulsion Laboratory",
     category: "Astrophysics",
     img_url: "https://res.cloudinary.com/debanftke/image/upload/v1648910315/AP-5_roi8hw.webp"
   },
   {
-    _id: 8,
+    _id: uuid(),
     title: "Astrobiology Winter 2022",
     creator: "UW Astrobiology",
     category: "Astrobiology",
     img_url: "https://res.cloudinary.com/debanftke/image/upload/v1648910896/AB-1_vnpygq.webp"
   },
   {
-    _id: 9,
+    _id: uuid(),
     title: "Amazing journey from Earth to end of the Universe",
     creator: "CoconutScienceLab",
     category: "Universe",
     img_url: "https://res.cloudinary.com/debanftke/image/upload/v1648912308/U-3_iwneq1.webp"
   },
   {
-    _id: 10,
+    _id: uuid(),
     title: "What is Astrophysics Explained",
     creator: "Insane Curiosity",
     category: "Astrophysics",
     img_url: "https://res.cloudinary.com/debanftke/image/upload/v1648909751/AP-4_bgeo8m.webp"
   },
   {
-    _id: 11,
+    _id: uuid(),
     title: "15 Years in Space: NASA's Apitzer Space Telescope",
     creator: "NASA Jet Propulsion Laboratory",
     category: "Astrophysics",
