+++ src/pages/index.js
@@ -2,3 +2,5 @@ export {History} from "./history"
 export {LikedVideos} from "./likedVideo"
 export {Playlist} from "./playlist"
 export {Watchlater} from "./watchlater"
+export {Home} from "./home"
+export {VideoList} from "./videolist"
